third_party/WebKit/LayoutTests/fast/events/before-unload-return-bad-value-expected.txt

CONSOLE ERROR: line 16: Uncaught Exception in toString()
Tests that an exception is thrown when the value returned in the beforeunload callback cannot be converted to a String

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
